+++ PersonTbc.Data/Models/Person.cs
@@ -6,7 +6,7 @@ public class Person : BaseModel<int>
 {
     [Required] public string FirstName { get; set; }
     [Required] public string LastName { get; set; }
-    [Required] public uint PersonalId { get; set; }
+    [Required] public ulong PersonalId { get; set; }
     public DateTime? DateOfBirth { get; set; }
 
     [Required] public int GenderId { get; set; }
